Async updates for designate and telemetry
18.0-FR1 async fixes for designate operator @ 390385c64c15d301639e610529c208ecd486c65e

18.0-FR1 async fixes for telemetry operator from upstream 18.0-fr1-async which cherry-picked only 3129d79 and a524c87

NOTE: infra-operator doesn't seem to have any struct changes for the
associated 'make serviceNetwork optional' Jira. I guess that struct must
not have been in sync for the FR1 release to begin with? Checked this
multiple times and don't think it is needed here but we do need to
release a new infra-operator anyway due to the labeling issue also
discovered.